Refactor routing, fix restart
