+++ etc/service/calico-vpp-agent/run
@@ -1,4 +1,4 @@
 #!/bin/sh
 exec 2>&1
-rm /var/run/calico/cni-server.sock
+rm -f /var/run/calico/cni-server.sock
 exec /bin/calico-vpp-agent
