+++ routing/connectivity/ipsec.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package routing
+package connectivity
 
 import (
 	"bytes"
@@ -25,22 +25,15 @@ import (
 	"github.com/calico-vpp/vpplink"
 	"github.com/calico-vpp/vpplink/types"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 )
 
-type ipsecProvider struct {
+type IpsecProvider struct {
+	*ConnectivityProviderData
 	ipipIfs map[string][]uint32
-	log     *logrus.Entry
-	s       *Server
 }
 
-func newIPsecProvider(s *Server) (p *ipsecProvider) {
-	p = &ipsecProvider{
-		ipipIfs: make(map[string][]uint32),
-		log:     s.log.WithField("connectivity", "ipsec"),
-		s:       s,
-	}
-	return p
+func NewIPsecProvider(d *ConnectivityProviderData) *IpsecProvider {
+	return &IpsecProvider{d, make(map[string][]uint32)}
 }
 
 func ipToSafeString(addr net.IP) string {
@@ -51,7 +44,7 @@ func profileName(srcNodeAddr, destNodeAddr net.IP) string {
 	return "pr_" + ipToSafeString(srcNodeAddr) + "_to_" + ipToSafeString(destNodeAddr)
 }
 
-func (p ipsecProvider) setupTunnelWithIds(i int, j int, destNodeAddr net.IP, nodeIP net.IP) (err error) {
+func (p IpsecProvider) setupTunnelWithIds(i int, j int, destNodeAddr net.IP, nodeIP net.IP) (err error) {
 	src := net.IP(append([]byte(nil), nodeIP.To4()...))
 	src[2] += byte(i)
 	dst := net.IP(append([]byte(nil), destNodeAddr.To4()...))
@@ -65,8 +58,8 @@ func (p ipsecProvider) setupTunnelWithIds(i int, j int, destNodeAddr net.IP, nod
 	return nil
 }
 
-func (p ipsecProvider) setupTunnels(destNodeAddr net.IP) (err error) {
-	nodeIP := p.s.getNodeIP(vpplink.IsIP6(destNodeAddr))
+func (p IpsecProvider) setupTunnels(destNodeAddr net.IP) (err error) {
+	nodeIP := p.getNodeIP(vpplink.IsIP6(destNodeAddr))
 	for i := 0; i < config.IpsecAddressCount; i++ {
 		if config.CrossIpsecTunnels {
 			for j := 0; j < config.IpsecAddressCount; j++ {
@@ -86,20 +79,20 @@ func (p ipsecProvider) setupTunnels(destNodeAddr net.IP) (err error) {
 	return nil
 }
 
-func (p ipsecProvider) setupOneTunnel(src, dst net.IP, psk string) (tunSwIfIndex uint32, err error) {
-	swIfIndex, err := p.s.vpp.AddIpipTunnel(src, dst, 0)
+func (p IpsecProvider) setupOneTunnel(src, dst net.IP, psk string) (tunSwIfIndex uint32, err error) {
+	swIfIndex, err := p.vpp.AddIpipTunnel(src, dst, 0)
 	if err != nil {
 		return 0, errors.Wrapf(err, "Error adding ipip tunnel %s -> %s", src.String(), dst.String())
 	}
 
-	err = p.s.vpp.InterfaceSetUnnumbered(swIfIndex, config.DataInterfaceSwIfIndex)
+	err = p.vpp.InterfaceSetUnnumbered(swIfIndex, config.DataInterfaceSwIfIndex)
 	if err != nil {
 		// TODO : delete tunnel
 		return 0, errors.Wrapf(err, "Error seting ipip tunnel %d unnumbered: %s", swIfIndex)
 	}
 
 	// Always enable GSO feature on IPIP tunnel, only a tiny negative effect on perf if GSO is not enabled on the taps
-	err = p.s.vpp.EnableGSOFeature(swIfIndex)
+	err = p.vpp.EnableGSOFeature(swIfIndex)
 	if err != nil {
 		// TODO : delete tunnel
 		return 0, errors.Wrapf(err, "Error enabling gso for ipip interface")
@@ -107,33 +100,33 @@ func (p ipsecProvider) setupOneTunnel(src, dst net.IP, psk string) (tunSwIfIndex
 
 	// Add and configure related IKE profile
 	profile := profileName(src, dst)
-	err = p.s.vpp.AddIKEv2Profile(profile)
+	err = p.vpp.AddIKEv2Profile(profile)
 	if err != nil {
 		return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 	}
 
-	err = p.s.vpp.SetIKEv2PSKAuth(profile, psk)
+	err = p.vpp.SetIKEv2PSKAuth(profile, psk)
 	if err != nil {
 		return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 	}
 
-	err = p.s.vpp.SetIKEv2LocalIDAddress(profile, src)
+	err = p.vpp.SetIKEv2LocalIDAddress(profile, src)
 	if err != nil {
 		return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 	}
 
-	err = p.s.vpp.SetIKEv2RemoteIDAddress(profile, dst)
+	err = p.vpp.SetIKEv2RemoteIDAddress(profile, dst)
 	if err != nil {
 		return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 	}
 
-	err = p.s.vpp.SetIKEv2PermissiveTrafficSelectors(profile)
+	err = p.vpp.SetIKEv2PermissiveTrafficSelectors(profile)
 	if err != nil {
 		return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 	}
 
 	p.log.Infof("IKE: Profile %s = swifindex %d", profile, swIfIndex)
-	err = p.s.vpp.SetIKEv2TunnelInterface(profile, swIfIndex)
+	err = p.vpp.SetIKEv2TunnelInterface(profile, swIfIndex)
 	if err != nil {
 		return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 	}
@@ -141,17 +134,17 @@ func (p ipsecProvider) setupOneTunnel(src, dst net.IP, psk string) (tunSwIfIndex
 	// Compare addresses lexicographically to select an initiator
 	if bytes.Compare(src.To4(), dst.To4()) > 0 {
 		p.log.Infof("IKE: Set responder %s->%s", src.String(), dst.String())
-		err = p.s.vpp.SetIKEv2Responder(profile, config.DataInterfaceSwIfIndex, dst)
+		err = p.vpp.SetIKEv2Responder(profile, config.DataInterfaceSwIfIndex, dst)
 		if err != nil {
 			return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 		}
 
-		err = p.s.vpp.SetIKEv2DefaultTransforms(profile)
+		err = p.vpp.SetIKEv2DefaultTransforms(profile)
 		if err != nil {
 			return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 		}
 
-		err = p.s.vpp.IKEv2Initiate(profile)
+		err = p.vpp.IKEv2Initiate(profile)
 		if err != nil {
 			return 0, errors.Wrapf(err, "error configuring IPsec tunnel from %s to %s", src.String(), dst.String())
 		}
@@ -163,10 +156,10 @@ func (p ipsecProvider) setupOneTunnel(src, dst net.IP, psk string) (tunSwIfIndex
 	return swIfIndex, nil
 }
 
-func (p *ipsecProvider) waitForIPsecSA(profile string, ipipInterface uint32) {
+func (p *IpsecProvider) waitForIPsecSA(profile string, ipipInterface uint32) {
 	for {
 		time.Sleep(time.Second)
-		iface, err := p.s.vpp.GetInterfaceDetails(ipipInterface)
+		iface, err := p.vpp.GetInterfaceDetails(ipipInterface)
 		if err != nil {
 			p.log.Errorf("Cannot get IPIP tunnel %d status", ipipInterface)
 			return
@@ -192,7 +185,7 @@ func getIPSecRoutePaths(swIfIndices []uint32) []types.RoutePath {
 	return paths
 }
 
-func (p ipsecProvider) addConnectivity(cn *NodeConnectivity) (err error) {
+func (p IpsecProvider) AddConnectivity(cn *NodeConnectivity) (err error) {
 	if _, found := p.ipipIfs[cn.NextHop.String()]; !found {
 		err = p.setupTunnels(cn.NextHop)
 		if err != nil {
@@ -201,7 +194,7 @@ func (p ipsecProvider) addConnectivity(cn *NodeConnectivity) (err error) {
 	}
 	swIfIndices := p.ipipIfs[cn.NextHop.String()]
 	p.log.Infof("IPSEC: ADD %s via %s [%v]", cn.Dst.String(), cn.NextHop.String(), swIfIndices)
-	e := p.s.vpp.RouteAdd(&types.Route{
+	e := p.vpp.RouteAdd(&types.Route{
 		Dst:   &cn.Dst,
 		Paths: getIPSecRoutePaths(swIfIndices),
 	})
@@ -212,13 +205,13 @@ func (p ipsecProvider) addConnectivity(cn *NodeConnectivity) (err error) {
 	return errors.Wrap(err, "Error configuring routes")
 }
 
-func (p ipsecProvider) delConnectivity(cn *NodeConnectivity) (err error) {
+func (p IpsecProvider) DelConnectivity(cn *NodeConnectivity) (err error) {
 	swIfIndices, found := p.ipipIfs[cn.NextHop.String()]
 	if !found {
 		return errors.Errorf("Deleting unknown ipip tunnel %s", cn.NextHop.String())
 	}
 	p.log.Infof("IPSEC: DEL %s via %s [%v]", cn.Dst.String(), cn.NextHop.String(), swIfIndices)
-	e := p.s.vpp.RouteDel(&types.Route{
+	e := p.vpp.RouteDel(&types.Route{
 		Dst:   &cn.Dst,
 		Paths: getIPSecRoutePaths(swIfIndices),
 	})
